@@ -1,4 +1,4 @@
-class CreateFaeRoles < ActiveRecord::Migration
+class CreateFaeRoles < ActiveRecord::Migration[4.2]
   def change
     create_table :fae_roles do |t|
       t.string :name
