@@ -1,4 +1,4 @@
-class CreateFaePages < ActiveRecord::Migration
+class CreateFaePages < ActiveRecord::Migration[4.2]
   def change
     create_table :fae_static_pages do |t|
       t.string :title
