@@ -1,4 +1,4 @@
-class CreateFaeTextFields < ActiveRecord::Migration
+class CreateFaeTextFields < ActiveRecord::Migration[4.2]
   def change
     create_table :fae_text_fields do |t|
       t.references :contentable, polymorphic: true, index: true
