@@ -13,8 +13,8 @@
 //= require jquery
 //= require jquery_ujs
 //= require jquery.remotipart
-//= require jquery.ui.sortable
-//= require jquery.ui.datepicker
+//= require jquery-ui/widgets/sortable
+//= require jquery-ui/widgets/datepicker
 //= require underscore
 //= require json2
 //= require judge
