@@ -167,27 +167,15 @@ Fae.navigation = {
   stickyHeaders: function(just_headers) {
     just_headers = FCH.setDefault(just_headers, false);
 
-    if(FCH.exists('.js-content-header')) {
-      var $header = $('.js-content-header');
-      var sidebar_top_offset = (parseInt( $header.outerHeight(), 10) + 30) + 'px';
-      $('#js-sidenav').css('padding-top', sidebar_top_offset );
+    var $header = $('.js-content-header');
+    var sidebar_top_offset = (parseInt( $header.outerHeight(), 10) + 30) + 'px';
+    $('#js-sidenav').css('padding-top', sidebar_top_offset );
 
-      $header.sticky({
-        placeholder: true,
-        perpetual_placeholder: true,
-        ignore_placeholder_offsets: true
-      });
-
-    // @depreciation - remove else block in 2.0
-    } else {
-      var $header = $('.main_content-header');
-      var sidebar_top_offset = (parseInt( $header.outerHeight(), 10) + 30) + 'px';
-      $('#js-sidenav').css('padding-top', sidebar_top_offset );
-
-      $('.main_content-header').sticky({
-        placeholder: true
-      });
-    }
+    $header.sticky({
+      placeholder: true,
+      perpetual_placeholder: true,
+      ignore_placeholder_offsets: true
+    });
 
     if (!just_headers) {
       $('#js-sidenav').sticky();
