@@ -84,7 +84,7 @@ module Fae
       @klass = klass_base.classify.constantize             # used as class reference in this controller
       @klass_singular = klass_base.singularize             # used in index views
       @klass_humanized = @klass_name.singularize.humanize  # used in index views
-      @index_path = '/' + params[:controller]              # used in form_header and form_buttons partials
+      @index_path = '/' + params[:controller]              # used in form_header partial
       @new_path = @index_path + '/new'                     # used in index_header partial
     end
 
