@@ -29,9 +29,9 @@ module Fae
           end
         end
       elsif item.class.columns_hash[attribute.to_s].present? && item.class.columns_hash[attribute.to_s].type == :boolean
-        content_tag(:td) do attr_toggle(item, attribute) end
+        content_tag(:td, fae_toggle(item, attribute))
       else
-        content_tag(:td) do col_name_or_image(item, attribute) end
+        content_tag(:td, col_name_or_image(item, attribute))
       end
     end
 
