@@ -7,7 +7,7 @@ module Fae
     has_many :users
 
     acts_as_list add_new_at: :bottom
-    default_scope { order('-position DESC') }
+    default_scope { order(Arel.sql('-position DESC')) }
 
     scope :public_roles, -> {where.not(name: 'super admin')}
 
