@@ -1,8 +1,8 @@
 # encoding: utf-8
 module Fae
   class ImageUploader < CarrierWave::Uploader::Base
-    include CarrierWave::MimeTypes
-    include CarrierWave::RMagick
+
+    include CarrierWave::MiniMagick
 
     # saves file size to DB
     process :save_file_size_in_model
