@@ -1,3 +1,3 @@
 module Fae
-  VERSION = '1.7.1'
+  VERSION = '2.0.0'
 end
