@@ -23,31 +23,20 @@ describe 'utilities#toggle' do
     it "shouldn't be able to toggle user's active attr" do
       create_super_user
 
-      if Rails::VERSION::MAJOR > 4
-        post '/admin/toggle/fae_users/1/active', as: :js
-        follow_redirect!
+      post '/admin/toggle/fae_users/1/active', as: :js
+      follow_redirect!
 
-        expect(response.body).to include(I18n.t(:devise)[:failure][:unauthenticated])
-      else
-        post '/admin/toggle/fae_users/1/active', format: :js
-        expect(response.status).to eq(401)
-      end
+      expect(response.body).to include(I18n.t(:devise)[:failure][:unauthenticated])
     end
 
     it "shouldn't be able to toggle release's on_prod attr" do
       create_super_user
       release = FactoryGirl.create(:release)
 
-      if Rails::VERSION::MAJOR > 4
-        post "/admin/toggle/releases/#{release.id}/on_prod", as: :js
-        follow_redirect!
+      post "/admin/toggle/releases/#{release.id}/on_prod", as: :js
+      follow_redirect!
 
-        expect(response.body).to include(I18n.t(:devise)[:failure][:unauthenticated])
-      else
-        post "/admin/toggle/releases/#{release.id}/on_prod", format: :js
-
-        expect(response.status).to eq(401)
-      end
+      expect(response.body).to include(I18n.t(:devise)[:failure][:unauthenticated])
     end
   end
 
