@@ -64,12 +64,10 @@ RSpec.configure do |config|
   # config.allow_url('gravatar.com')
 
   # https://github.com/rails/rails-controller-testing#rspec
-  if Rails::VERSION::MAJOR > 4
-    [:controller, :view, :request].each do |type|
-      config.include ::Rails::Controller::Testing::TestProcess, :type => type
-      config.include ::Rails::Controller::Testing::TemplateAssertions, :type => type
-      config.include ::Rails::Controller::Testing::Integration, :type => type
-    end
+  [:controller, :view, :request].each do |type|
+    config.include ::Rails::Controller::Testing::TestProcess, :type => type
+    config.include ::Rails::Controller::Testing::TemplateAssertions, :type => type
+    config.include ::Rails::Controller::Testing::Integration, :type => type
   end
 end
 
